1st Commit
As of now the cImage file is being worked on, the function is supposed to create the negative rgb of an image and does work to an extend. Currently, it changes the values but is stuck as an array. I'm currently trying to change the array to an Abstract image again.